bit more memory saving

+++ bl3d/utils.py
@@ -136,7 +136,7 @@ def prob2labels(detection, segmentation_, seg_threshold=0.5, min_voxels=65,
     peaks[~binary_masks] = False # restrict to peaks in cell bodies
     masks = morphology.watershed(-segmentation_, morphology.label(peaks).astype(np.int32),
                                  mask=binary_masks, connectivity=3,
-                                 compactness=compactness_factor)
+                                 compactness=compactness_factor).astype(np.int32)
     print(masks.max(), 'initial cells')
 
     # Remove masks that are too small or too big (usually bad detections)
@@ -144,6 +144,7 @@ def prob2labels(detection, segmentation_, seg_threshold=0.5, min_voxels=65,
     to_keep = np.logical_and(mask_sizes >= min_voxels, mask_sizes <= max_voxels)
     masks[~to_keep[masks]] = 0  # set to background
     label, _, _ = segmentation.relabel_sequential(masks)
+    label = label.astype(np.int32)
     print(label.max(), 'final cells')
 
     return label
